refactor: mypage 폴더 도메인과 컴포넌트 분리
- 기존 app/ 폴더 안에 있던 mypage 하위 폴더를 app 바깥으로 이동
- app/mypage/ 폴더는 page.tsx만 존재
- mypage 하위 폴더에서 언더바 삭제
- import 경로 수정
