+++ src/app/(full-layout)/mypage/page.tsx
@@ -1,0 +1,5 @@
+import MypageCard from '@/mypage/components/mypage-card/mypage-card';
+
+export default function MypagePage() {
+  return <MypageCard />;
+}
