--- src/app/mypage/page.tsx
@@ -1,5 +1,0 @@
-import MypageCard from './_components/mypage-card/mypage-card';
-
-export default function MypagePage() {
-  return <MypageCard />;
-}
